web/phisher2/files/phisher2/requirements.txt
pyocr == 0.8.3
selenium == 4.9.1
Pillow == 9.5.0
flask == 2.3.2
requests == 2.31.0
uwsgi == 2.0.21